make row_test,blob_test,one_test stand-alone tests
