+++ test/integration_test.dart
@@ -1,27 +1,21 @@
 library integrationtests;
 
 import 'dart:async';
-import 'dart:typed_data';
 
 import 'package:logging/logging.dart';
 import 'package:options_file/options_file.dart';
-import 'package:sqljocky/constants.dart';
 import 'package:sqljocky/sqljocky.dart';
-import 'package:sqljocky/utils.dart';
 import 'package:test/test.dart';
 
 import 'test_util.dart';
 
-part 'integration/blob.dart';
 part 'integration/charset.dart';
 part 'integration/errors.dart';
 part 'integration/execute_multi.dart';
 part 'integration/largeblob.dart';
 part 'integration/nullmap.dart';
 part 'integration/numbers.dart';
-part 'integration/one.dart';
 part 'integration/prepared_query.dart';
-part 'integration/row.dart';
 part 'integration/stored_procedures.dart';
 part 'integration/stream.dart';
 part 'integration/two.dart';
@@ -51,15 +45,12 @@ void main() {
   var host = options.getString('host', 'localhost');
 
   runPreparedQueryTests(user, password, db, port, host);
-  runIntTests(user, password, db, port, host);
   runIntTests2(user, password, db, port, host);
   runCharsetTests(user, password, db, port, host);
   runNullMapTests(user, password, db, port, host);
   runNumberTests(user, password, db, port, host);
   runStreamTests(user, password, db, port, host);
-  runRowTests(user, password, db, port, host);
   runErrorTests(user, password, db, port, host);
-  runBlobTests(user, password, db, port, host);
 //  runStoredProcedureTests(user, password, db, port, host);
   runExecuteMultiTests(user, password, db, port, host);
 //  if (results['large_packets'] == 'true') {
